feat: added download section to preview-eml
