+++ src/types/utils/index.d.ts
@@ -136,3 +136,9 @@ export type ThemeObj = {
 	>;
 	avatarColors: Record<`avatar_${number}`, string>;
 };
+
+export type GetAttachmentsDownloadLinkProps = {
+	messageId: string;
+	messageSubject: string;
+	attachments: Array<string | undefined>;
+};
