+++ translations/en.json
@@ -238,6 +238,9 @@
         "attachment_plural": "Attachments",
         "attachment_status": "Status of e-mail item",
         "attachment_type": "Attachment type",
+        "attachment_with_count": "{{count}} Attachment",
+        "attachment_with_count_plural": "{{count}} Attachments",
+        "attachments_disclaimer": "You are viewing an attached message. The authenticity of the attached messages can not be verified.",
         "automatically_delete_duplicates": "Automatically delete duplicate copies of the same message when received",
         "back": "Back",
         "bcc": "Bcc",
